Allow using key & configuration parser from tincd in tinc.
